refactor-settings * Move useUpdateManifest to its own file
